mam22-floa to double and more

+++ maman22/mycomp/command_parse.h
@@ -34,7 +34,7 @@ int read_var_name(char *input, int *index, char *place_to_token, char *var);
  * Function get place to token for internal use
  * Side effect: increment the index to next token index,
  * as I explained in get_next_token documentation*/
-int read_number(char *input, int *index, char *place_to_token, float *var);
+int read_number(char *input, int *index, char *place_to_token, double *var);
 
 /*function get pointer to char array (input) and offset (index)
 * and validate string ends starting from index.
@@ -49,10 +49,10 @@ int read_end_of_string(char *input, int *index, char *place_to_token);
  * function try to read from input, starting from index, what the format reqired.
  * Format can contain 3 characters:
  * c - represents variable name
- * f - represents float number
+ * f - represents double number
  * , - represent comma
  * For example, read_in_format(input,index,place_to_token,"c,f,f",&c,&f1,&f2)
- * try to read var name to c, and after it float nuber to f1 and float number to f2,
+ * try to read var name to c, and after it double nuber to f1 and double number to f2,
  * separated by commas.
  * If it success, it returns 1, otherwise, it returns 0 and print error message to user.
  * Function get place to token for internal use
